Edit Comment Done

+++ Services/CommentService.cs
@@ -1,42 +1,46 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using OmniTalks.Contracs;
+using OmniTalks.Controllers;
 using OmniTalks.Data;
 using OmniTalks.Models;
 using OmniTalks.Models.Domein;
+using System.CodeDom;
 
 namespace OmniTalks.Services
 {
-    public class CommentService : ICommentService
-    {
-        public ApplicationDbContext _context;
+	public class CommentService : ICommentService
+	{
+		public ApplicationDbContext _context;
 
-        public CommentService(ApplicationDbContext context)
-        {
-            _context = context;
-        }
+		public CommentService(ApplicationDbContext context)
+		{
+			_context = context;
+		}
 
-        public async Task Add(AddCommentViewModel model, string id)
-        {
-            Comment comment = new Comment()
-            {
-                Id = Guid.NewGuid(),
-                Text = model.Text,
-                PostId = model.PostId,
-                UserId = new Guid(id),
-            };
+		public async Task Add(AddCommentViewModel model, string id)
+		{
+			Comment comment = new Comment()
+			{
+				Id = Guid.NewGuid(),
+				Text = model.Text,
+				PostId = model.PostId,
+				UserId = new Guid(id),
+			};
 
-            await _context.Comments.AddAsync(comment);
-            await _context.SaveChangesAsync();
-        }
+			await _context.Comments.AddAsync(comment);
+			await _context.SaveChangesAsync();
+		}
 
-        public async Task Remove(Guid id)
-        {
-            var comment = await _context.Comments
-                .FirstOrDefaultAsync(x => x.Id == id);
+		public async Task Remove(Guid id, Guid userId)
+		{
+			Comment comment = await this._context.Comments.FindAsync(id);
+			if (comment.UserId == userId)
+			{
+				_context.Comments.Remove(comment);
+				await _context.SaveChangesAsync();
+			}
 
-            _context.Comments.Remove(comment);
-            await _context.SaveChangesAsync();
-        }
+		}
 
 		public async Task AddLike(CommentLikeViewModel model, string userId)
 		{
@@ -65,10 +69,25 @@ namespace OmniTalks.Services
 
 		public async Task Edit(CommentViewModel model)
 		{
-			Comment comment = await _context.Comments.FindAsync(model.Id);
-			comment.Text = model.Text;
+			Comment comment = await this._context.Comments.FindAsync(model.Id);
+			comment.Text = model.Text.ToString();
 
 			await _context.SaveChangesAsync();
+		}
+
+		public async Task<CommentViewModel> Rewrite(Guid id)
+		{
+			var comment = await _context.Comments.FindAsync(id);
+			CommentViewModel model = new CommentViewModel();
+			model.Text = comment.Text;
+
+			return model;
+		}
+
+		public async Task<Guid> GetById(Guid id)
+		{
+			var comment = await _context.Comments.FindAsync(id);
+			return comment.UserId;
 		}
 	}
 }
